+++ lib/demoflu.dart
@@ -1,9 +1,10 @@
 library demoflu;
 
 export 'src/app.dart';
+export 'src/macro.dart' hide MacroFactoryHelper;
 export 'src/menu_item.dart';
 export 'src/theme.dart';
-export 'src/page/page.dart';
+export 'src/page/page.dart' hide DemoFluPageHelper;
 export 'src/page/page_section_group.dart';
 export 'src/page/banner_section.dart';
 export 'src/page/console_section.dart';