+++ lib/src/widgets/page_widget.dart
@@ -1,3 +1,4 @@
+import 'package:demoflu/src/macro.dart';
 import 'package:demoflu/src/menu_item.dart';
 import 'package:demoflu/src/page/page.dart';
 import 'package:demoflu/src/model.dart';
@@ -20,6 +21,9 @@ class DemoFluPageWidget extends StatelessWidget {
     final DemoMenuItem menuItem = model.selectedMenuItem;
     if (menuItem.page != null) {
       final DemoFluPage page = menuItem.page!();
+      MacroFactory macroFactory = DemoFluProvider.macroFactoryOf(context);
+      DemoFluPageHelper.setMacroFactory(page, macroFactory);
+      page.initialize(context);
       children.add(BreadcrumbWidget(menuItem: menuItem));
       children.add(SizedBox(height: 24));
       children.add(SectionCollectionWidget(collection: page));